Moved modules/topics tests from subfolders into main folders.
